bridge/requirements.txt
google-genai>=1.0.0
telnyx>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=13.0
numpy
scipy
python-dotenv
httpx
